+++ example/android/app/src/main/kotlin/com/example/example/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.example
+package com.nt4f04unds_widgets.example
 
 import io.flutter.embedding.android.FlutterActivity