+++ lib/src/constants.dart
@@ -3,8 +3,6 @@
 *  Licensed under the BSD-style license. See LICENSE in the project root for license information.
 *--------------------------------------------------------------------------------------------*/
 
-import 'package:flutter/material.dart';
-
 class NFConstants {
   /// Default icon size I prefer to use.
   static const double iconSize = 25.0;
@@ -14,10 +12,4 @@ class NFConstants {
 
   /// Default app bar height I prefer to use
   static const double toolbarHeight = 52.0;
-
-  /// Locales that library supports
-  static const List<Locale> supportedLocales = [
-    Locale('en', 'US'),
-    Locale('ru', 'RU'),
-  ];
 }