+++ lib/src/core/gestures/drag.dart
@@ -11,8 +11,6 @@
 /// TODO: migrate this back to null safety (copited this on flutter master 43e1d1596e)
 /// ###########################################################################################
 
-import 'dart:ui';
-
 import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
@@ -81,10 +79,10 @@ abstract class NFDragGestureRecognizer extends OneSequenceGestureRecognizer {
   /// {@macro flutter.gestures.GestureRecognizer.kind}
   NFDragGestureRecognizer({
     Object? debugOwner,
-    PointerDeviceKind? kind,
+    Set<PointerDeviceKind>? supportedDevices,
     this.dragStartBehavior = DragStartBehavior.start,
     this.velocityTrackerBuilder = _defaultBuilder,
-  }) : super(debugOwner: debugOwner, kind: kind);
+  }) : super(debugOwner: debugOwner, supportedDevices: supportedDevices);
 
   /// Called on each pointer move event (even before the drag was accepted).
   ///
@@ -544,8 +542,8 @@ class NFVerticalDragGestureRecognizer extends NFDragGestureRecognizer {
   /// {@macro flutter.gestures.GestureRecognizer.kind}
   NFVerticalDragGestureRecognizer({
     Object? debugOwner,
-    PointerDeviceKind? kind,
-  }) : super(debugOwner: debugOwner, kind: kind);
+    Set<PointerDeviceKind>? supportedDevices,
+  }) : super(debugOwner: debugOwner, supportedDevices: supportedDevices);
 
   @override
   bool isFlingGesture(VelocityEstimate estimate, PointerDeviceKind kind) {
@@ -589,8 +587,8 @@ class NFHorizontalDragGestureRecognizer extends NFDragGestureRecognizer {
   /// {@macro flutter.gestures.GestureRecognizer.kind}
   NFHorizontalDragGestureRecognizer({
     Object? debugOwner,
-    PointerDeviceKind? kind,
-  }) : super(debugOwner: debugOwner, kind: kind);
+    Set<PointerDeviceKind>? supportedDevices,
+  }) : super(debugOwner: debugOwner, supportedDevices: supportedDevices);
 
   @override
   bool isFlingGesture(VelocityEstimate estimate, PointerDeviceKind kind) {