+++ lib/src/core/routes/routes.dart
@@ -3,6 +3,5 @@
 *  Licensed under the BSD-style license. See LICENSE in the project root for license information.
 *--------------------------------------------------------------------------------------------*/
 
-export 'back_button_listener.dart';
 export 'dismissible_route.dart';
 export 'route_transitions.dart';